+++ tests/utils/test_version.py
@@ -1,6 +1,6 @@
 import pytest
 
-from openeo.capabilities import ApiVersionException, ComparableVersion
+from openeo.utils.version import ApiVersionException, ComparableVersion
 
 
 class TestComparableVersion:
@@ -18,11 +18,14 @@ class TestComparableVersion:
         assert ComparableVersion(ComparableVersion("1.2.3")).parts == (1, 2, 3)
         assert ComparableVersion(ComparableVersion("1.b.3")).parts == (1, "b", 3)
 
-    @pytest.mark.parametrize(["a", "b", "c"], [
-        (ComparableVersion("1.2.3"), ComparableVersion("1.2.3"), ComparableVersion("2.3.4")),
-        (ComparableVersion("1.2.3"), "1.2.3", "2.3.4"),
-        ("1.2.3", ComparableVersion("1.2.3"), ComparableVersion("2.3.4")),
-    ])
+    @pytest.mark.parametrize(
+        ["a", "b", "c"],
+        [
+            (ComparableVersion("1.2.3"), ComparableVersion("1.2.3"), ComparableVersion("2.3.4")),
+            (ComparableVersion("1.2.3"), "1.2.3", "2.3.4"),
+            ("1.2.3", ComparableVersion("1.2.3"), ComparableVersion("2.3.4")),
+        ],
+    )
     def test_equals(self, a, b, c):
         assert (a == b) is True
         assert (a == c) is False
@@ -60,10 +63,16 @@ class TestComparableVersion:
         assert ComparableVersion(str(ComparableVersion("1.2.3"))).parts == (1, 2, 3)
         assert ComparableVersion(str(ComparableVersion("1.b.3"))).parts == (1, "b", 3)
 
-    @pytest.mark.parametrize("b", [
-        "0.9", "1", "1.2.2",
-        ComparableVersion("0.9"), ComparableVersion("1.1"),
-    ])
+    @pytest.mark.parametrize(
+        "b",
+        [
+            "0.9",
+            "1",
+            "1.2.2",
+            ComparableVersion("0.9"),
+            ComparableVersion("1.1"),
+        ],
+    )
     def test_operators(self, b):
         a = ComparableVersion("1.2.3")
         assert (a == a) is True
@@ -78,48 +87,48 @@ class TestComparableVersion:
         assert (b >= a) is False
 
     def test_right_referencing(self):
-        v = ComparableVersion('1.2.3')
-        assert v.equals('1.2.3')
-        assert v.above('0')
-        assert v.above('0.1')
-        assert v.above('0.1.2')
-        assert v.above('1.2')
-        assert v.above('1.2.2')
-        assert v.above('1.2.2b')
-        assert v.above('1.2.3') is False
-        assert v.above('1.2.20') is False
-        assert v.above('1.2.4') is False
-        assert v.above('1.10.4') is False
-        assert v.at_least('0')
-        assert v.at_least('1')
-        assert v.at_least('1.1')
-        assert v.at_least('1.10') is False
-        assert v.at_least('1.2')
-        assert v.at_least('1.02')
-        assert v.at_least('1.2.2')
-        assert v.at_least('1.2.3')
-        assert v.at_least('1.2.3a') is False
-        assert v.at_least('1.2.4') is False
-        assert v.at_least('1.3') is False
-        assert v.at_least('2') is False
-        assert v.below('2')
-        assert v.below('1.3')
-        assert v.below('1.2.4')
-        assert v.below('1.2.3b')
-        assert v.below('1.2.3') is False
-        assert v.below('1.2') is False
-        assert v.at_most('2')
-        assert v.at_most('1.3')
-        assert v.at_most('1.2.3c')
-        assert v.at_most('1.2.3')
-        assert v.at_most('1.02.03')
-        assert v.at_most('1.2.2b') is False
-        assert v.at_most('1.2') is False
-        assert v.at_most('1.10')
+        v = ComparableVersion("1.2.3")
+        assert v.equals("1.2.3")
+        assert v.above("0")
+        assert v.above("0.1")
+        assert v.above("0.1.2")
+        assert v.above("1.2")
+        assert v.above("1.2.2")
+        assert v.above("1.2.2b")
+        assert v.above("1.2.3") is False
+        assert v.above("1.2.20") is False
+        assert v.above("1.2.4") is False
+        assert v.above("1.10.4") is False
+        assert v.at_least("0")
+        assert v.at_least("1")
+        assert v.at_least("1.1")
+        assert v.at_least("1.10") is False
+        assert v.at_least("1.2")
+        assert v.at_least("1.02")
+        assert v.at_least("1.2.2")
+        assert v.at_least("1.2.3")
+        assert v.at_least("1.2.3a") is False
+        assert v.at_least("1.2.4") is False
+        assert v.at_least("1.3") is False
+        assert v.at_least("2") is False
+        assert v.below("2")
+        assert v.below("1.3")
+        assert v.below("1.2.4")
+        assert v.below("1.2.3b")
+        assert v.below("1.2.3") is False
+        assert v.below("1.2") is False
+        assert v.at_most("2")
+        assert v.at_most("1.3")
+        assert v.at_most("1.2.3c")
+        assert v.at_most("1.2.3")
+        assert v.at_most("1.02.03")
+        assert v.at_most("1.2.2b") is False
+        assert v.at_most("1.2") is False
+        assert v.at_most("1.10")
 
-        assert v.above(ComparableVersion('1.2'))
-        assert v.at_least(ComparableVersion('1.2.3a')) is False
-        assert v.at_most(ComparableVersion('1.02.03'))
+        assert v.above(ComparableVersion("1.2"))
+        assert v.at_least(ComparableVersion("1.2.3a")) is False
+        assert v.at_most(ComparableVersion("1.02.03"))
 
     def test_left_referencing(self):
         v = ComparableVersion("1.2.3")